refactor: Gathering 관련 enum 클래스 정리  - core.domain.gathering.enums 패키지에서 보관
