+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/gathering/GatheringPreview.java
@@ -1,5 +1,8 @@
 package com.triplem.momoim.core.domain.gathering;
 
+import com.triplem.momoim.core.domain.gathering.enums.GatheringLocation;
+import com.triplem.momoim.core.domain.gathering.enums.GatheringStatus;
+import com.triplem.momoim.core.domain.gathering.enums.GatheringType;
 import com.triplem.momoim.core.domain.member.GatheringMemberDetail;
 import java.time.LocalDateTime;
 import java.util.List;
