+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/gathering/enums/GatheringSortType.java
@@ -1,4 +1,4 @@
-package com.triplem.momoim.core.domain.gathering;
+package com.triplem.momoim.core.domain.gathering.enums;
 
 import lombok.Getter;
