+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/gathering/enums/GatheringType.java
@@ -1,4 +1,4 @@
-package com.triplem.momoim.core.domain.gathering;
+package com.triplem.momoim.core.domain.gathering.enums;
 
 public enum GatheringType {
     OFFLINE(), ZOOM(), DISCORD(), GOOGLE_MEET(), KAKAO_TALK(), ETC()
