+++ momoim-external-api/src/main/java/com/triplem/momoim/api/gathering/request/RegisterGatheringRequest.java
@@ -2,11 +2,11 @@ package com.triplem.momoim.api.gathering.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.triplem.momoim.core.domain.gathering.Gathering;
-import com.triplem.momoim.core.domain.gathering.GatheringCategory;
-import com.triplem.momoim.core.domain.gathering.GatheringLocation;
-import com.triplem.momoim.core.domain.gathering.GatheringStatus;
-import com.triplem.momoim.core.domain.gathering.GatheringSubCategory;
-import com.triplem.momoim.core.domain.gathering.GatheringType;
+import com.triplem.momoim.core.domain.gathering.enums.GatheringCategory;
+import com.triplem.momoim.core.domain.gathering.enums.GatheringLocation;
+import com.triplem.momoim.core.domain.gathering.enums.GatheringStatus;
+import com.triplem.momoim.core.domain.gathering.enums.GatheringSubCategory;
+import com.triplem.momoim.core.domain.gathering.enums.GatheringType;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.Getter;
